requirements.txt
streamlit>=1.30.0
pandas>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.2
plotly>=5.17.0
openai>=1.0.0
pydantic>=2.0.0 